# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vectric Blog Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List blog posts (public, paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 12 }
        - name: category
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [published, draft, scheduled] }
        - name: search
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Paginated posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/trending:
    get:
      operationId: getTrendingPosts
      tags: [posts]
      summary: Get trending posts (by view count)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Trending posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostSummary"
  /posts/featured:
    get:
      operationId: getFeaturedPosts
      tags: [posts]
      summary: Get featured/hero posts
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 3 }
      responses:
        "200":
          description: Featured posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostSummary"
  /posts/recent:
    get:
      operationId: getRecentPosts
      tags: [posts]
      summary: Get most recent published posts
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Recent posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostSummary"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found

    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /posts/slug/{slug}:
    get:
      operationId: getPostBySlug
      tags: [posts]
      summary: Get post by SEO slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
  /posts/{id}/view:
    post:
      operationId: incrementPostView
      tags: [posts]
      summary: Increment view count for a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: View recorded
  /posts/{id}/related:
    get:
      operationId: listRelatedPosts
      tags: [posts]
      summary: Get related posts (returns up to 3)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostSummary"

  # --- CATEGORIES ---
...
components:
  schemas:
...
    PostSummary:
      type: object
      required: [id, title, slug, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: ["string", "null"] }
        featuredImageUrl: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        authorName: { type: ["string", "null"] }
        readingTime: { type: ["integer", "null"] }
        viewCount: { type: integer }
        status: { type: string, enum: [published, draft, scheduled] }
        isFeatured: { type: boolean }
        publishedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        tags: { type: array, items: { type: string } }
    Post:
      type: object
      required: [id, title, slug, content, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: ["string", "null"] }
        content: { type: string }
        featuredImageUrl: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        authorName: { type: ["string", "null"] }
        readingTime: { type: ["integer", "null"] }
        viewCount: { type: integer }
        status: { type: string, enum: [published, draft, scheduled] }
        isFeatured: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        seoKeywords: { type: ["string", "null"] }
        ctaButtons: { type: ["string", "null"] }
        scheduledAt: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        tags: { type: array, items: { type: string } }
    PostInput:
      type: object
      required: [title, content]
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        featuredImageUrl: { type: string }
        categoryId: { type: integer }
        authorName: { type: string }
        readingTime: { type: integer }
        status: { type: string, enum: [published, draft, scheduled] }
        isFeatured: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoKeywords: { type: string }
        ctaButtons: { type: string }
        scheduledAt: { type: string }
        tags: { type: array, items: { type: string } }
    PostUpdate:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        featuredImageUrl: { type: string }
        categoryId: { type: integer }
        authorName: { type: string }
        readingTime: { type: integer }
        status: { type: string, enum: [published, draft, scheduled] }
        isFeatured: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoKeywords: { type: string }
        ctaButtons: { type: string }
        scheduledAt: { type: string }
        tags: { type: array, items: { type: string } }
    PostList:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/PostSummary"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # CATEGORIES
...